Update as of 2020-10-24 20:43PM: 1. Updated data type from decimal to double from model MembershipType. 2. Added Initial Migration. 3. Updated Program.cs file. 4. Updated Startup.cs file.

+++ Vidly/Models/MembershipType.cs
@@ -13,11 +13,11 @@ namespace Vidly.Models
 
         public string Name { get; set; }
 
-        public decimal SignUpFee { get; set; }
+        public double SignUpFee { get; set; }
 
         public int DurationInMonths { get; set; }
 
-        public decimal DiscountRate { get; set; }
+        public double DiscountRate { get; set; }
 
     }
 }
